+++ CHANGELOG.md
@@ -212,3 +212,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Sending not indexed updates
 - Force flag on get_by_id request
 - Custom new from/to support
+
+## [3.2.3] - 2017-11-07 16:04:00
+### Added
+- Batches, debugging api
+- Timezone to vers request
+### Fixed
+- Special price from/to support